+++ packages/lib-core/src/index.ts
@@ -43,7 +43,6 @@ export { useFeatureFlag, UseFeatureFlagResult } from './runtime/useFeatureFlag';
 
 // Core utilities
 export { applyCodeRefSymbol } from './runtime/coderefs';
-export { isLocalPluginManifest, isStandardPluginManifest } from './runtime/plugin-manifest';
 
 // Testing utilities
 export { TestPluginStore } from './testing/TestPluginStore';
@@ -67,11 +66,10 @@ export {
 export { ResourceFetch } from './types/fetch';
 export { PluginLoadResult, PluginLoaderInterface } from './types/loader';
 export {
-  PluginRegistrationMethod,
   PluginRuntimeMetadata,
-  PluginManifest,
+  RemotePluginManifest,
   LocalPluginManifest,
-  AnyPluginManifest,
+  PluginManifest,
   PendingPlugin,
   LoadedPlugin,
   FailedPlugin,